pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lightning-toll"
version = "0.1.0"
description = "L402 Lightning paywalls for FastAPI — monetize any API endpoint with Bitcoin Lightning"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Jeletor"}]
keywords = ["lightning", "bitcoin", "l402", "paywall", "fastapi", "api", "micropayments"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "websockets>=11.0",
    "httpx>=0.24.0",
    "coincurve>=18.0",
    "pycryptodome>=3.19",
]

[project.optional-dependencies]
fastapi = ["fastapi>=0.100.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "httpx>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/jeletor/lightning-toll-python"
Repository = "https://github.com/jeletor/lightning-toll-python"
Issues = "https://github.com/jeletor/lightning-toll-python/issues"

[tool.setuptools.packages.find]
include = ["lightning_toll*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
